# backend/openapi.yaml
openapi: 3.1.0
info:
  title: PodChat Backend API
  version: 1.0.0
  description: |
    REST API for PodChat — live podcast platform with rooms, recordings,
    LiveKit audio and Socket.IO realtime presence.
servers:
  - url: http://localhost:3001/api
    description: Local dev
  - url: https://livepodchat.com/api
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code: { type: string }
        requiresPassword: { type: boolean }
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string, format: email }
        avatarUrl: { type: string, format: uri }
        bio: { type: string }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user: { $ref: '#/components/schemas/User' }
        token: { type: string, description: JWT }
    Room:
      type: object
      required: [id, slug, title, hostId, status, maxSpeakers, isPublic, hasPassword, createdAt]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        hostId: { type: string, format: uuid }
        status: { type: string, enum: [waiting, live, ended] }
        maxSpeakers: { type: integer }
        isPublic: { type: boolean }
        hasPassword: { type: boolean }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
    PublicRoom:
      type: object
      required: [id, slug, title, status, isPublic, hasPassword, host, participantCount, maxSpeakers, createdAt]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        status: { type: string, enum: [waiting, live] }
        isPublic: { type: boolean }
        hasPassword: { type: boolean }
        host:
          type: object
          properties:
            id: { type: string }
            username: { type: string }
            avatarUrl: { type: string }
        participantCount: { type: integer }
        maxSpeakers: { type: integer }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [auth]
      security: []
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [auth]
      summary: Current authenticated user
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /rooms:
    get:
      tags: [rooms]
      summary: List public rooms (waiting + live)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms: { type: array, items: { $ref: '#/components/schemas/PublicRoom' } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
    post:
      tags: [rooms]
      summary: Create a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 100 }
                isPublic: { type: boolean, default: true }
                password: { type: string, description: Optional — makes the room password-protected }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Room' } } } }
...